+++ src/frontend/app/features/applications/applications.routing.ts
@@ -26,10 +26,10 @@ import { EditApplicationComponent } from './edit-application/edit-application.co
 import { AddRouteStepperComponent } from './routes/add-route-stepper/add-route-stepper.component';
 import { SshApplicationComponent } from './ssh-application/ssh-application.component';
 import { DynamicExtenstionRoutes } from '../../core/extension/dynamic-extension-routes';
-import { StratosActionType, StratosTabType } from '../../core/extension/extension-service';
+import { StratosActionType, StratosTabType, extensionsActionRouteKey } from '../../core/extension/extension-service';
 import { PageNotFoundComponentComponent } from '../../core/page-not-found-component/page-not-found-component.component';
 
-const appplicationsRoutes: Routes = [
+const applicationsRoutes: Routes = [
   {
     path: 'new',
     component: CreateApplicationComponent,
@@ -45,6 +45,9 @@ const appplicationsRoutes: Routes = [
         path: '',
         component: ApplicationWallComponent,
         pathMatch: 'full',
+        data: {
+          extensionsActionsKey: StratosActionType.Applications
+        }
       },
       {
         path: ':cfId/:id',
@@ -74,7 +77,8 @@ const appplicationsRoutes: Routes = [
             path: '',
             component: ApplicationTabsBaseComponent,
             data: {
-              uiFullView: true
+              uiFullView: true,
+              extensionsActionsKey: StratosActionType.Application
             },
             children: [
               { path: '', redirectTo: 'summary', pathMatch: 'full' },
@@ -87,7 +91,8 @@ const appplicationsRoutes: Routes = [
               { path: 'events', component: EventsTabComponent },
               { path: 'github', component: GithubTabComponent },
               { path: 'metrics', component: MetricsTabComponent },
-              { path: '**',
+              {
+                path: '**',
                 component: PageNotFoundComponentComponent,
                 canActivate: [DynamicExtenstionRoutes],
                 data: {
@@ -100,7 +105,8 @@ const appplicationsRoutes: Routes = [
             path: 'add-route',
             component: AddRouteStepperComponent,
           },
-          { path: '**',
+          {
+            path: '**',
             component: PageNotFoundComponentComponent,
             canActivate: [DynamicExtenstionRoutes],
             data: {
@@ -111,7 +117,8 @@ const appplicationsRoutes: Routes = [
       }
     ]
   },
-  { path: '**',
+  {
+    path: '**',
     component: PageNotFoundComponentComponent,
     canActivate: [DynamicExtenstionRoutes],
     data: {
@@ -124,7 +131,7 @@ const appplicationsRoutes: Routes = [
   imports: [
     CreateApplicationModule,
     DeployApplicationModule,
-    RouterModule.forChild(appplicationsRoutes)
+    RouterModule.forChild(applicationsRoutes)
 
   ]
 })