+++ src/frontend/app/shared/components/boolean-indicator/boolean-indicator.component.ts
@@ -1,12 +1,12 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { split } from 'ts-node/dist';
+import { Component, Input } from '@angular/core';
 
 export enum BooleanIndicatorType {
   enabledDisabled = 'enabled-disabled',
   lockedUnlocked = 'locked-unlocked',
   unlockedLocked = 'unlocked-locked',
   yesNo = 'yes-no',
-  trueFalse = 'true-false'
+  trueFalse = 'true-false',
+  healthyUnhealthy = 'healthy-unhealthy'
 }
 
 
@@ -15,38 +15,41 @@ export enum BooleanIndicatorType {
   templateUrl: './boolean-indicator.component.html',
   styleUrls: ['./boolean-indicator.component.scss']
 })
-export class BooleanIndicatorComponent implements OnInit {
+export class BooleanIndicatorComponent {
 
   @Input() isTrue: boolean;
   @Input() type: BooleanIndicatorType;
+  @Input() showText = true;
 
-  // Should we use a subtle display - this won't show the No option as dandger (typically red)
+  // Should we use a subtle display - this won't show the No option as danger (typically red)
   @Input() subtle = true;
+
+  // Invert the text labels with the icons (No text for yes value and vice-versa)
+  @Input() inverse = false;
 
   private icons = {
     Yes: 'check_circle',
     Enabled: 'check_circle',
+    Healthy: 'check_circle',
     True: 'check_circle',
     Add: 'add_circle',
     No: 'highlight_off',
     Disabled: 'highlight_off',
+    Unhealthy: 'highlight_off',
     False: 'highlight_off',
     Remove: 'remove_circle',
     Locked: 'lock_outline',
     Unlocked: 'lock_open',
   };
-  constructor() { }
-
-  ngOnInit() {
-  }
 
   getIcon = () => {
-    return this.icons[this.getText()];
+    return this.icons[this.getText(this.inverse)];
   }
 
-  getText = (): string => {
+  getText = (inverse = false): string => {
     const [enabledText, disabledText] = this.getTypeText(this.type);
-    return this.capitalizeFirstLetter(this.isTrue ? enabledText : disabledText);
+    const value = inverse ? !this.isTrue : this.isTrue;
+    return this.capitalizeFirstLetter(value ? enabledText : disabledText);
   }
 
   getTypeText = (s) => s.split('-');