+++ src/test-e2e/cloud-foundry/cf-level/cf-top-level-e2e.spec.ts
@@ -2,7 +2,9 @@ import { e2e, E2ESetup } from '../../e2e';
 import { E2EConfigCloudFoundry } from '../../e2e.types';
 import { ConsoleUserType } from '../../helpers/e2e-helpers';
 import { CfTopLevelPage } from './cf-top-level-page.po';
-
+import { SideNavMenuItem } from '../../po/side-nav.po';
+import { CFPage } from '../../po/cf-page.po';
+import { ListComponent } from '../../po/list.po';
 
 describe('CF - Top Level - ', () => {
 
@@ -10,40 +12,40 @@ describe('CF - Top Level - ', () => {
   let e2eSetup: E2ESetup;
   let defaultCf: E2EConfigCloudFoundry;
 
-  function setup(user: ConsoleUserType) {
+  function navToCfPage() {
+    // There is only one CF endpoint registered (since that is what we setup)
+    const page = new CFPage();
+    page.sideNav.goto(SideNavMenuItem.CloudFoundry);
+    CfTopLevelPage.detect().then(p => {
+      cfPage = p;
+      cfPage.waitForPageOrChildPage();
+      cfPage.loadingIndicator.waitUntilNotShown();
+    });
+  }
+
+  beforeAll(() => {
+    defaultCf = e2e.secrets.getDefaultCFEndpoint();
     e2eSetup = e2e.setup(ConsoleUserType.admin)
       .clearAllEndpoints()
       .registerDefaultCloudFoundry()
-      .connectAllEndpoints(user)
-      .loginAs(user)
-      .getInfo(user);
-  }
-
-
-  function navToCfPage() {
-    defaultCf = e2e.secrets.getDefaultCFEndpoint();
-    const endpointGuid = e2e.helper.getEndpointGuid(e2e.info, defaultCf.name);
-    cfPage = CfTopLevelPage.forEndpoint(endpointGuid);
-    cfPage.navigateTo();
-    cfPage.waitForPageOrChildPage();
-    cfPage.loadingIndicator.waitUntilNotShown();
-  }
-
-  function testTitle() {
-    expect(cfPage.header.getTitleText()).toBe(defaultCf.name);
-  }
-
+      .connectAllEndpoints(ConsoleUserType.admin)
+      .connectAllEndpoints(ConsoleUserType.user);
+  });
 
   describe('As Admin', () => {
-    beforeEach(() => {
-      setup(ConsoleUserType.admin);
+    beforeAll(() => {
+      e2eSetup.loginAs(ConsoleUserType.admin);
     });
 
     describe('Basic Tests -', () => {
-
       beforeEach(navToCfPage);
 
-      it('Breadcrumb', testTitle);
+      beforeEach(() => {
+      });
+
+      it('Breadcrumb', () => {
+        expect(cfPage.header.getTitleText()).toBe(defaultCf.name);
+      });
 
       it('Summary Panel', () => {
         expect(cfPage.waitForInstanceAddress().getValue()).toBe(defaultCf.url);
@@ -67,8 +69,9 @@ describe('CF - Top Level - ', () => {
   });
 
   describe('As User', () => {
-    beforeEach(() => {
-      setup(ConsoleUserType.user);
+    beforeAll(() => {
+      e2eSetup = e2e.setup(ConsoleUserType.admin)
+        .loginAs(ConsoleUserType.user);
     });
 
     describe('Basic Tests -', () => {
@@ -87,15 +90,20 @@ describe('CF - Top Level - ', () => {
 
       it('Walk Tabs', () => {
         cfPage.goToOrgTab();
-        cfPage.goToUsersTab();
-        // cfPage.goToFirehoseTab();// Is not shown to non-admins
-        cfPage.goToFeatureFlagsTab();
-        cfPage.goToBuildPacksTab();
-        cfPage.goToStacksTab();
-        cfPage.goToSecurityGroupsTab();
-        cfPage.goToSummaryTab();
+        const orgs = new ListComponent();
+        orgs.waitUntilShown();
+        orgs.getTotalResults().then(totalOrgs => {
+          if (totalOrgs <= 12) {
+            cfPage.goToUsersTab();
+          }
+          // cfPage.goToFirehoseTab();// Is not shown to non-admins
+          cfPage.goToFeatureFlagsTab();
+          cfPage.goToBuildPacksTab();
+          cfPage.goToStacksTab();
+          cfPage.goToSecurityGroupsTab();
+          cfPage.goToSummaryTab();
+        });
       });
-
     });
   });