+++ src/test-e2e/cloud-foundry/organizations-spaces-e2e.spec.ts
@@ -12,13 +12,13 @@ import { CfTopLevelPage } from './cf-level/cf-top-level-page.po';
 describe('CF - Manage Organizations and Spaces', () => {
 
   const testOrgName = e2e.helper.getCustomerOrgSpaceLabel(null, 'org');
+  const testOrg2Name = e2e.helper.getCustomerOrgSpaceLabel(null, 'org2');
   const testSpaceName = e2e.helper.getCustomerOrgSpaceLabel(null, 'space');
   let endpointGuid;
 
   let cloudFoundry: CfTopLevelPage;
 
   let cfHelper: CFHelpers;
-  const listComponent = new ListComponent();
 
   beforeAll(() => {
     const setup = e2e.setup(ConsoleUserType.admin)
@@ -44,7 +44,10 @@ describe('CF - Manage Organizations and Spaces', () => {
     cloudFoundry.waitForPageOrChildPage();
   });
 
-  afterAll(() => cfHelper.deleteOrgIfExisting(endpointGuid, testOrgName));
+  afterAll(() => Promise.all([
+    cfHelper.deleteOrgIfExisting(endpointGuid, testOrgName),
+    cfHelper.deleteOrgIfExisting(endpointGuid, testOrg2Name)
+  ]));
 
   it('Should validate org name', () => {
     const cardView = cloudFoundry.goToOrgView();
@@ -91,10 +94,10 @@ describe('CF - Manage Organizations and Spaces', () => {
     cardView.cards.waitUntilShown();
 
     // Delete the org
-    cardView.cards.waitForCardByTitle(testOrgName).then(card => {
+    cardView.cards.findCardByTitle(testOrgName, MetaCardTitleType.CUSTOM, true).then(card => {
       card.openActionMenu().then(menu => {
         menu.clickItem('Delete');
-        ConfirmDialogComponent.expectDialogAndConfirm('Delete', 'Delete Organization');
+        ConfirmDialogComponent.expectDialogAndConfirm('Delete', 'Delete Organization', testOrgName);
         // Wait until the card has gone
         card.waitUntilNotShown();
       });
@@ -128,7 +131,7 @@ describe('CF - Manage Organizations and Spaces', () => {
     const cardView = cloudFoundry.goToOrgView();
     const list = new ListComponent();
     list.refresh();
-    cardView.cards.findCardByTitle(testOrgName).then(org => {
+    cardView.cards.findCardByTitle(testOrgName, MetaCardTitleType.CUSTOM, true).then(org => {
       org.click();
 
       cloudFoundry.subHeader.clickItem('Spaces');
@@ -150,10 +153,10 @@ describe('CF - Manage Organizations and Spaces', () => {
       cardView.cards.waitUntilShown();
 
       // Get the card for the space
-      cardView.cards.findCardByTitle(testSpaceName).then(space => {
+      cardView.cards.findCardByTitle(testSpaceName).then((space: MetaCard) => {
         space.openActionMenu().then(menu => {
           menu.clickItem('Delete');
-          ConfirmDialogComponent.expectDialogAndConfirm('Delete', 'Delete Space');
+          ConfirmDialogComponent.expectDialogAndConfirm('Delete', 'Delete Space', testSpaceName);
           cardView.cards.getCardCount().then(c => {
             expect(c).toBe(0);
           });
@@ -161,5 +164,43 @@ describe('CF - Manage Organizations and Spaces', () => {
       });
     });
   });
-});
 
+  it('Should create an org and a space', () => {
+    const cardView = cloudFoundry.goToOrgView();
+
+    // Click the add button to add an organization
+    cloudFoundry.header.clickIconButton('add');
+    const modal = new StepperComponent();
+    modal.getStepperForm().fill({
+      'orgname': testOrg2Name
+    });
+    modal.waitUntilCanNext('Create');
+    modal.next();
+
+    cardView.cards.waitUntilShown();
+
+    // Go to the org and create a space
+    cardView.cards.findCardByTitle(testOrg2Name, MetaCardTitleType.CUSTOM, true).then(org => {
+      org.click();
+
+      cloudFoundry.subHeader.clickItem('Spaces');
+      cardView.cards.waitUntilShown();
+      const list = new ListComponent();
+      list.refresh();
+
+      // Add space
+      // Click the add button to add a space
+      cloudFoundry.header.clickIconButton('add');
+
+      modal.getStepperForm().fill({
+        'spacename': testSpaceName
+      });
+      expect(modal.canNext()).toBeTruthy();
+      modal.next();
+
+      cloudFoundry.subHeader.clickItem('Spaces');
+      cardView.cards.waitUntilShown();
+    });
+
+  });
+});